Improve performance, SEO metadata, and maintainability tooling
