+++ package.json
@@ -6,7 +6,12 @@
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "typecheck": "tsc --noEmit",
+    "format:check": "npx --yes prettier@3.6.2 --check .",
+    "test": "node --test tests/**/*.test.mjs",
+    "test:e2e:smoke": "node scripts/e2e-smoke.mjs",
+    "ci": "npm run lint && npm run typecheck && npm run format:check && npm run test && npm run test:e2e:smoke"
   },
   "dependencies": {
     "framer-motion": "^12.23.3",
